feat: Allow keeping unknown files in export/ dir
- Configuration option _ExportKeepFileNames_ added
- Updated .env.example and appsettings.json

+++ NanoBlog/Extensions/DirectoryInfoExtensions.cs
@@ -59,17 +59,31 @@ public static class DirectoryInfoExtensions
         return fileStream;
     }
 
-    public static void Clear(this DirectoryInfo directoryInfo)
+    public static void Clear(
+        this DirectoryInfo directoryInfo,
+        ICollection<string> keepNames,
+        StringComparer nameStringComparer
+    )
     {
         directoryInfo.Refresh();
 
         foreach (var fileInfo in directoryInfo.EnumerateFiles())
         {
+            if (keepNames?.Contains(fileInfo.Name, nameStringComparer) == true)
+            {
+                continue;
+            }
+
             fileInfo.Delete();
         }
 
         foreach (var subDirectoryInfo in directoryInfo.EnumerateDirectories())
         {
+            if (keepNames?.Contains(subDirectoryInfo.Name, nameStringComparer) == true)
+            {
+                continue;
+            }
+
             Directory.Delete(subDirectoryInfo.FullName, true);
         }
     }
